requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
Django==5.0.4
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.15.1
drf-yasg==1.21.7
html5lib==1.1
idna==3.7
inflection==0.5.1
lxml==5.2.1
oscrypto==1.3.0
packaging==24.0
pdf2image==1.17.0
pdfkit==1.0.0
pillow==10.3.0
pycparser==2.22
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
